Use global variable
